Clean up after aosp/3652453
Fix a few rounding errors that were causing test failures

Remove Float.toFloat() (and a few other Type.toType())

s/round().toInt()/roundToInt() - they handle ties differently

Make Point/Rect equality-with-tolerance checks in PdfViewPaginationTest
produce more readable error messages.

Use larger tolerances where needed to make tests pass, without allowing for more than 1 pixel difference between expect and actual.

Test: ./gradlew :pdf:pdf-viewer:cAT same for integration tests

+++ pdf/pdf-viewer/src/main/kotlin/androidx/pdf/view/Page.kt
@@ -252,7 +252,7 @@ internal class Page(
             // Highlight locations are defined in content coordinates, compute their location
             // in View coordinates using locationInView
             highlightRect.set(highlight.area.pageRect)
-            highlightRect.offset(locationInView.left.toFloat(), locationInView.top.toFloat())
+            highlightRect.offset(locationInView.left, locationInView.top)
             highlightPaint.color = highlight.color
             canvas.drawRect(highlightRect, highlightPaint)
         }
@@ -263,10 +263,7 @@ internal class Page(
                 ?.filter { it.widgetType == FormWidgetInfo.WIDGET_TYPE_TEXTFIELD }
                 ?.forEach {
                     formWidgetHighlightRect.set(it.widgetRect)
-                    formWidgetHighlightRect.offset(
-                        locationInView.left.toFloat(),
-                        locationInView.top.toFloat(),
-                    )
+                    formWidgetHighlightRect.offset(locationInView.left, locationInView.top)
                     highlightPaint.color = pdfFormFillingConfig.formFieldsHighlightColor
                     canvas.drawRect(formWidgetHighlightRect, highlightPaint)
                 }
